gst-plugins-camera-s5p6818: autotools compiling
This patch changes the way of the compiling this library using autotools.
The previous 'Makefile' is renamed to 'Makefile.cross'
and the library needs a extra dependency path for gstreamer.
so, the path('gst_path') is added using with option when we set a configuration.
(--with-gst_path=the dir where the libraries exist about gstreamer)

Change-Id: Iaa11cc972f56cc4b9b8e5e71dc1c8bec8c52006e
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>